version text moved on config page #132
